requirements.txt
asgiref==3.7.2

beautifulsoup4==4.12.3
bootpay-backend==2.0.8

certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5

decouple~=0.0.7
defusedxml==0.7.1
Django==5.0.2
django-allauth==0.61.1
django-cron==0.6.0
django-crontab==0.7.1
django-user-agents==0.4.0
djangorestframework==3.14.0

elasticsearch==7.13.2

html5lib==1.1

idna==3.6

joblib==1.4.2

lxml==5.2.2
langchain-openai==0.2.1

mysqlclient==2.2.4

numpy==1.26.4

oauthlib==3.2.2
openai==1.75.0
olefile==0.47
opencv-python==4.11.0.86

pandas==2.2.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyPDF2==3.0.1
pywin32==310
pyhwpx==0.51.2
pytesseract==0.3.13
paddlepaddle==3.0.0
paddleocr==2.10.0
pdfplumber==0.11.6
python-decouple==3.8

requests==2.31.0
requests-oauthlib==1.3.1

scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4

threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.13.2
tzdata==2024.1

ua-parser==0.18.0
urllib3==1.26.20
user-agents==2.2.0

webencodings==0.5.1



